Shahar: moved folder & final
